backend logger system started, controllers finished.

+++ backend/controllers/userController.js
@@ -26,6 +26,8 @@ const signUpUser = async (req, res) => {
         typeof email !== "string" ||
         typeof password !== "string"
     ) {
+        logger.warn("Signup rejected: Invalid signup data: firstname, lastname, username, email, password !== string");
+
         return res.status(400).json({
             error: "Invalid signup data.",
         });
@@ -38,6 +40,8 @@ const signUpUser = async (req, res) => {
         !email.trim() ||
         !password.trim()
     ) {
+        logger.warn("Signup rejected: on of required fields is empty");
+
         return res.status(400).json({
             error: "All signup fields are required.",
         });
@@ -55,18 +59,21 @@ const signUpUser = async (req, res) => {
         ]);
 
         if (usernameExists) {
+            logger.warn("Signup rejected: duplicate account data: username");
             return res.status(400).json({
                 error: "Username is already in use.",
             });
         }
 
         if (emailExists) {
+            logger.warn("Signup rejected: duplicate account data: email");
             return res.status(400).json({
                 error: "Email is already in use.",
             });
         }
 
         if (password.length < 8) {
+            logger.warn("Signup rejected: bad password");
             return res.status(400).json({
                 error: "Password is not strong enough.",
             });
@@ -83,6 +90,12 @@ const signUpUser = async (req, res) => {
             password: hashedPass
         });
         const token = createToken(user._id);
+
+        logger.info({
+            username: normalizedUsername,
+            userId: user._id.toString(),
+        }, "Signup successful");
+
         res.status(201).json({
             user: {
                 _id: user._id,
@@ -99,13 +112,16 @@ const signUpUser = async (req, res) => {
         });
 
     } catch (error) {
-        console.error("Signup error:", error);
-
         if (error.code === 11000) {
+            logger.warn("Signup rejected: duplicate account data");
             return res.status(409).json({
                 error: "Username or email is already in use.",
             });
         }
+
+        logger.error({
+            err: error,
+        }, "Signup failed");
 
         return res.status(500).json({
             error: "Failed to create account.",
@@ -123,6 +139,7 @@ const signInUser = async (req, res) => {
             !email.trim() ||
             !password
         ) {
+            logger.warn(" email or password no provided");
             return res.status(400).json({
                 error: "Email and password are required.",
             });
@@ -133,6 +150,9 @@ const signInUser = async (req, res) => {
         const user = await User.findOne({ email: normalizedEmail });
 
         if (!user) {
+            logger.warn({
+                email: normalizedEmail,
+            }, " Invalid credentials: !user on database");
             return res.status(401).json({
                 error: "Invalid credentials.",
             });
@@ -141,12 +161,17 @@ const signInUser = async (req, res) => {
         const isPassCorrect = await bcrypt.compare(password, user.password);
 
         if (!isPassCorrect) {
+            logger.warn(" Invalid credentials: wrong pass");
             return res.status(401).json({
                 error: "Invalid credentials.",
             });
         }
 
         const token = createToken(user._id);
+        logger.info({
+            email: normalizedEmail,
+            userId: user._id.toString(),
+        }, "Signin successful");
 
         res.status(200).json({
             user: {
@@ -164,8 +189,9 @@ const signInUser = async (req, res) => {
         });
 
     } catch (error) {
-        console.error("Signin error:", error);
-
+        logger.error({
+            err: error,
+        }, "Signup failed");
         return res.status(500).json({
             error: "Failed to sign in.",
         });
@@ -173,19 +199,31 @@ const signInUser = async (req, res) => {
 };
 
 const getUser = async (req, res) => {
+
     try {
         const id = req.user._id;
         if (!mongoose.isValidObjectId(id)) {
+            logger.warn({
+                userId: id.toString(),
+            }, "id is not a valid object");
             return res.status(404).json({ message: "User not found!" });
         };
         const user = await User.findById(id).select("-password").populate("favoriteToilets").populate("reviews");
-        user ? (
-            res.status(200).json(user)
-        ) : (
-            res.status(404).json({ message: "User not found!" })
-        )
+        if (!user) {
+            logger.warn({
+                userId: id.toString(),
+            }, "id is not a valid object");
+            return res.status(404).json({ message: "User not found!" })
+        };
+        logger.info({
+            userId: user._id.toString(),
+        }, "Get user successful");
+        return res.status(200).json(user)
+
     } catch (error) {
-        console.error("Get user error:", error);
+        logger.error({
+            err: error,
+        }, "Get user failed");
         return res.status(500).json({ message: "Failed to load user", });
     }
 };
@@ -193,16 +231,28 @@ const getUser = async (req, res) => {
 const removeUser = async (req, res) => {
     try {
         const id = req.user._id;
+        if (!mongoose.isValidObjectId(id)) {
+            logger.warn({
+                userId: id.toString(),
+            }, "id is not a valid object");
+            return res.status(404).json({ message: "User not found!" });
+        };
+
         const deletedUser = await User.findByIdAndDelete(id);
 
         if (!deletedUser) {
+            logger.warn("user not found");
             return res.status(404).json({ error: "User not found." })
         }
-
+        logger.warn({
+            userId: id.toString(),
+        }, "user delet successful");
         res.status(204).send();
 
     } catch (error) {
-        console.error("Get user error:", error);
+        logger.error({
+            err: error,
+        }, "delete user failed");
         res.status(500).json({ error: "Failed to delete user." })
     }
